Move utility functions into test-utils
