+++ test/api3-server-v1/AirseekerRegistry.sol.ts
@@ -1,125 +1,18 @@
 import type { HardhatEthersSigner } from '@nomicfoundation/hardhat-ethers/signers';
 import * as helpers from '@nomicfoundation/hardhat-network-helpers';
 import { expect } from 'chai';
-import type { AddressLike, BigNumberish, BytesLike, HDNodeWallet } from 'ethers';
+import type { HDNodeWallet } from 'ethers';
 import { ethers } from 'hardhat';
 
-import type { AirseekerRegistry, Api3ServerV1 } from '../../src/index';
-
-export async function updateBeacon(
-  api3ServerV1: Api3ServerV1,
-  feedName: string,
-  airnode: HDNodeWallet,
-  timestamp: BigNumberish,
-  value: BigNumberish
-) {
-  const encodedValue = ethers.AbiCoder.defaultAbiCoder().encode(['int224'], [value]);
-  const templateId = deriveTemplateId(`OIS title of Airnode with address ${airnode.address}`, feedName);
-  const beaconId = deriveBeaconId(airnode.address, templateId);
-  await api3ServerV1.updateBeaconWithSignedData(
-    airnode.address,
-    templateId,
-    timestamp,
-    encodedValue,
-    await airnode.signMessage(
-      ethers.toBeArray(
-        ethers.solidityPackedKeccak256(['bytes32', 'uint256', 'bytes'], [templateId, timestamp, encodedValue])
-      )
-    )
-  );
-  return {
-    templateId,
-    beaconId,
-  };
-}
-
-export async function updateBeaconSet(
-  api3ServerV1: Api3ServerV1,
-  feedName: string,
-  airnodes: HDNodeWallet[],
-  timestamp: BigNumberish,
-  value: BigNumberish
-) {
-  const encodedValue = ethers.AbiCoder.defaultAbiCoder().encode(['int224'], [value]);
-
-  const beaconUpdateData = airnodes.map((airnode) => {
-    const templateId = deriveTemplateId(`OIS title of Airnode with address ${airnode.address}`, feedName);
-    return {
-      airnode,
-      templateId,
-      beaconId: deriveBeaconId(airnode.address, templateId),
-    };
-  });
-  for (const beaconUpdateDatum of beaconUpdateData) {
-    await api3ServerV1.updateBeaconWithSignedData(
-      beaconUpdateDatum.airnode.address,
-      beaconUpdateDatum.templateId,
-      timestamp,
-      encodedValue,
-      await beaconUpdateDatum.airnode.signMessage(
-        ethers.toBeArray(
-          ethers.solidityPackedKeccak256(
-            ['bytes32', 'uint256', 'bytes'],
-            [beaconUpdateDatum.templateId, timestamp, encodedValue]
-          )
-        )
-      )
-    );
-  }
-  const beaconIds = beaconUpdateData.map((beaconUpdateDatum) => beaconUpdateDatum.beaconId);
-  await api3ServerV1.updateBeaconSetWithBeacons(beaconIds);
-  return {
-    templateIds: beaconUpdateData.map((beaconUpdateDatum) => beaconUpdateDatum.templateId),
-    beaconIds,
-    beaconSetId: deriveBeaconSetId(beaconIds),
-  };
-}
-
-export async function readBeacons(api3ServerV1: Api3ServerV1, beaconIds: BytesLike[]) {
-  const returndata = await api3ServerV1.multicall.staticCall(
-    beaconIds.map((beaconId) => api3ServerV1.interface.encodeFunctionData('dataFeeds', [beaconId]))
-  );
-  return returndata
-    .map((returndata) => ethers.AbiCoder.defaultAbiCoder().decode(['int224', 'uint32'], returndata))
-    .map((decodedReturnData) => {
-      return { value: decodedReturnData[0], timestamp: decodedReturnData[1] };
-    });
-}
-
-export function encodeUpdateParameters(
-  deviationThreshold: number,
-  deviationReference: number,
-  heartbeatInterval: number
-) {
-  return ethers.AbiCoder.defaultAbiCoder().encode(
-    ['uint256', 'int224', 'uint256'],
-    [deviationThreshold, deviationReference, heartbeatInterval]
-  );
-}
-
-function deriveTemplateId(oisTitle: string, feedName: string) {
-  const endpointId = ethers.solidityPackedKeccak256(['string', 'string'], [oisTitle, 'feed']);
-  // Parameters encoded in Airnode ABI
-  // https://docs.api3.org/reference/airnode/latest/specifications/airnode-abi.html
-  return ethers.solidityPackedKeccak256(
-    ['bytes32', 'bytes'],
-    [
-      endpointId,
-      ethers.AbiCoder.defaultAbiCoder().encode(
-        ['bytes32', 'bytes32', 'bytes32'],
-        [ethers.encodeBytes32String('1b'), ethers.encodeBytes32String('name'), ethers.encodeBytes32String(feedName)]
-      ),
-    ]
-  );
-}
-
-function deriveBeaconId(airnodeAddress: AddressLike, templateId: BytesLike) {
-  return ethers.solidityPackedKeccak256(['address', 'bytes32'], [airnodeAddress, templateId]);
-}
-
-function deriveBeaconSetId(beaconIds: BytesLike[]) {
-  return ethers.keccak256(ethers.AbiCoder.defaultAbiCoder().encode(['bytes32[]'], [beaconIds]));
-}
+import type { AirseekerRegistry } from '../../src/index';
+import {
+  deriveBeaconId,
+  deriveBeaconSetId,
+  deriveTemplateId,
+  encodeUpdateParameters,
+  readBeacons,
+  updateBeaconSet,
+} from '../test-utils';
 
 async function registerBeaconSet(airseekerRegistry: AirseekerRegistry, feedName: string, airnodes: HDNodeWallet[]) {
   const beacons = airnodes
